refactor styles registry
