+++ app/styling/styled-jsx/registry.tsx
@@ -1,19 +1,19 @@
 'use client';
 
-import React from 'react';
+import React, { useState } from 'react';
 import { useServerInsertedHTML } from 'next/navigation';
-import { useStyledJsxRegistry } from '#/lib/styling';
+import { StyleRegistry, createStyleRegistry } from 'styled-jsx';
 
-export default function StyledJsxRegistry({
-  children,
-}: {
-  children: React.ReactNode;
-}) {
-  const [StyledJsxRegistry, styledJsxFlushEffect] = useStyledJsxRegistry();
+type ChildProps = { children: React.ReactNode };
+
+export default function StyledJsxRegistry({ children }: ChildProps) {
+  const [jsxStyleRegistry] = useState(() => createStyleRegistry());
 
   useServerInsertedHTML(() => {
-    return <>{styledJsxFlushEffect()}</>;
+    const styles = jsxStyleRegistry.styles();
+    jsxStyleRegistry.flush();
+    return <>{styles}</>;
   });
 
-  return <StyledJsxRegistry>{children}</StyledJsxRegistry>;
+  return <StyleRegistry registry={jsxStyleRegistry}>{children}</StyleRegistry>;
 }
